+++ src/app/modules/market/component/market-exchange/market-exchange.component.ts
@@ -4,7 +4,7 @@ import { TradeFetchResultEntry, TradeFetchService } from '@shared/module/poe/tra
 import { TradeExchangeRequest, TradeExchangeResponse, TradeExchangeService } from '@shared/module/poe/trade/exchange';
 import { BehaviorSubject, Observable, of, Subject } from 'rxjs';
 import { catchError, flatMap, map, switchMap, takeUntil, tap } from 'rxjs/operators';
-import { SendEvent } from '../market-exchange-price/market-exchange-price.component';
+import { WhisperEvent } from '../market-exchange-price/market-exchange-price.component';
 
 const DEFAULT_REQUEST = () => {
   const request: TradeExchangeRequest = {
@@ -99,7 +99,7 @@ export class MarketExchangeComponent implements OnInit, OnDestroy {
     this.clear();
   }
 
-  public onSend(whisper: string, event: SendEvent): void {
+  public onWhisper(whisper: string, event: WhisperEvent): void {
     let copy = whisper.slice();
     copy = copy.replace('{0}', `${event.itemAmount}`);
     copy = copy.replace('{1}', `${event.valueAmount}`);