+++ src/app/modules/market/component/market-item-search/market-item-search.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, OnDestroy, OnInit, Output, TemplateRef, ViewChild } from '@angular/core';
 import { ChatService } from '@shared/module/poe/chat';
-import { TradeFetchEntryListing, TradeFetchResultEntry, TradeFetchService, TradeSearchRequest, TradeSearchResponse, TradeSearchService } from '@shared/module/poe/trade';
+import { TradeFetchResultEntry, TradeFetchService, TradeSearchRequest, TradeSearchResponse, TradeSearchService } from '@shared/module/poe/trade';
 import { BehaviorSubject, Observable, of, Subject } from 'rxjs';
 import { catchError, flatMap, map, switchMap, takeUntil, tap } from 'rxjs/operators';
 
@@ -106,8 +106,8 @@ export class MarketItemSearchComponent implements OnInit, OnDestroy {
     this.clear();
   }
 
-  public onSend(listing: TradeFetchEntryListing): void {
-    this.chat.send(listing.whisper);
+  public onWhisper(whisper: string): void {
+    this.chat.send(whisper);
   }
 
   private clear(): void {