+++ src/app/modules/trade/component/trade-settings/trade-settings.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { AudioFile, AudioService } from '@app/audio';
+import { AudioService } from '@app/audio';
 import { EnumValues } from '@app/enum';
 import { FeatureSettingsComponent } from '@app/feature';
 import { TradeFeatureSettings, TradeFilter, TradeLayout } from '@modules/trade/trade-feature-settings';
@@ -27,6 +27,21 @@ export class TradeSettingsComponent extends FeatureSettingsComponent<TradeFeatur
   }
 
   public onPlay(): void {
-    this.audio.play(AudioFile.Notification, this.settings.tradeSoundVolume / 100);
+    this.audio.play(this.settings.tradeSound, this.settings.tradeSoundVolume / 100);
+  }
+
+  public onSelect(input: HTMLInputElement): void {
+    input.click();
+  }
+
+  public onSelected(input: HTMLInputElement): void {
+    if (input.files.length > 0) {
+      const reader = new FileReader();
+      reader.onload = (e) => {
+        this.settings.tradeSound = e.target.result as string;
+        this.save();
+      };
+      reader.readAsDataURL(input.files[0]);
+    }
   }
 }